Refactoring and windows support upd
Significant code modifications were performed in the database and file management systems: datatype enumerations were renamed for clarity, logic for mapping and unmapping memory in files was enhanced. Numerous updates were also made to balance memory usage and improve code performance.

+++ src/backend/journal/varchar_mgr.c
@@ -20,7 +20,7 @@ int64_t vch_init(void){
 vch_ticket_t vch_add(int64_t vachar_mgr_idx, char* varchar){
     vch_ticket_t ticket;
     ticket.block = lb_alloc(vachar_mgr_idx);
-    ticket.size = strlen(varchar)+1;
+    ticket.size = (int64_t)strlen(varchar)+1;
     lb_write(
             vachar_mgr_idx,
             &ticket.block,
